+++ cli/replicaset/lua/cartridge/edit_replicasets_body.lua
@@ -1,0 +1,8 @@
+local cartridge = require('cartridge')
+local replicasets = ...
+
+local res, err = cartridge.admin_edit_topology({
+    replicasets = replicasets
+})
+
+assert(res, tostring(err))